Extract out the EditEngine
While trying this interface I had to make a few methods public which I
am not a fan of, will eventually circle back and fix them
